Change game to basically generate a grid of a certain size, move all other logic into grid
