+++ game_of_life/partial_solutions/60_percent/spec/game_of_life/grid_spec.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-RSpec.describe GameOfLife::Game do
-  subject(:game) { described_class.new }
+RSpec.describe GameOfLife::Grid do
+  subject(:grid) { described_class.new }
 
-  let(:max_index) { game.grid_size - 1 }
+  let(:max_index) { grid.grid_size - 1 }
   let(:alive_cell) { instance_double(GameOfLife::Cell, alive?: true, dead?: false) }
   let(:dead_cell) { instance_double(GameOfLife::Cell, alive?: false, dead?: true) }
